Added new \Common\Core\Form class to use inheritance for Backend and Frontend.
