+++ src/Frontend/Core/Engine/Form.php
@@ -10,9 +10,7 @@ namespace Frontend\Core\Engine;
  */
 
 use Symfony\Component\Filesystem\Filesystem;
-
-use Frontend\Core\Engine\Language AS FL;
-
+use Frontend\Core\Engine\Language as FL;
 
 /**
  * This is our extended version of SpoonForm.
@@ -20,22 +18,8 @@ use Frontend\Core\Engine\Language AS FL;
  * @author Davy Hellemans <davy.hellemans@netlash.com>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  */
-class Form extends \SpoonForm
+class Form extends \Common\Core\Form
 {
-    /**
-     * The header instance
-     *
-     * @var    Header
-     */
-    private $header;
-
-    /**
-     * The URL instance
-     *
-     * @var    Url
-     */
-    private $URL;
-
     /**
      * @param string $name     Name of the form.
      * @param string $action   The action (URL) whereto the form will be submitted, if not provided it will
@@ -99,26 +83,6 @@ class Form extends \SpoonForm
 
         // create and return a button
         return parent::addButton($name, $value, $type, $class);
-    }
-
-    /**
-     * Adds a single checkbox.
-     *
-     * @param string $name       The name of the element.
-     * @param bool   $checked    Should the checkbox be checked?
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @return \SpoonFormCheckbox
-     */
-    public function addCheckbox($name, $checked = false, $class = null, $classError = null)
-    {
-        $name = (string) $name;
-        $checked = (bool) $checked;
-        $class = ($class !== null) ? (string) $class : 'inputCheckbox';
-        $classError = ($classError !== null) ? (string) $classError : 'inputCheckboxError';
-
-        // create and return a checkbox
-        return parent::addCheckbox($name, $checked, $class, $classError);
     }
 
     /**
@@ -215,42 +179,6 @@ class Form extends \SpoonForm
     }
 
     /**
-     * Adds a single dropdown.
-     *
-     * @param string $name              Name of the element.
-     * @param array  $values            Values for the dropdown.
-     * @param string $selected          The selected elements.
-     * @param bool   $multipleSelection Is it possible to select multiple items?
-     * @param string $class             Class(es) that will be applied on the element.
-     * @param string $classError        Class(es) that will be applied on the element when an error occurs.
-     * @return \SpoonFormDropdown
-     */
-    public function addDropdown(
-        $name,
-        array $values = null,
-        $selected = null,
-        $multipleSelection = false,
-        $class = null,
-        $classError = null
-    ) {
-        $name = (string) $name;
-        $values = (array) $values;
-        $selected = ($selected !== null) ? $selected : null;
-        $multipleSelection = (bool) $multipleSelection;
-        $class = ($class !== null) ? (string) $class : 'select';
-        $classError = ($classError !== null) ? (string) $classError : 'selectError';
-
-        // special classes for multiple
-        if ($multipleSelection) {
-            $class .= ' selectMultiple';
-            $classError .= ' selectMultipleError';
-        }
-
-        // create and return a dropdown
-        return parent::addDropdown($name, $values, $selected, $multipleSelection, $class, $classError);
-    }
-
-    /**
      * Adds a single file field.
      *
      * @param string $name       Name of the element.
@@ -286,146 +214,6 @@ class Form extends \SpoonForm
         $this->add(new FrontendFormImage($name, $class, $classError));
 
         return $this->getField($name);
-    }
-
-    /**
-     * Adds a multiple checkbox.
-     *
-     * @param string $name       The name of the element.
-     * @param array  $values     The values for the checkboxes.
-     * @param mixed  $checked    Should the checkboxes be checked?
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @return \SpoonFormMultiCheckbox
-     */
-    public function addMultiCheckbox($name, array $values, $checked = null, $class = null, $classError = null)
-    {
-        $name = (string) $name;
-        $values = (array) $values;
-        $checked = ($checked !== null) ? (array) $checked : null;
-        $class = ($class !== null) ? (string) $class : 'inputCheckbox';
-        $classError = ($classError !== null) ? (string) $classError : 'inputCheckboxError';
-
-        // create and return a multi checkbox
-        return parent::addMultiCheckbox($name, $values, $checked, $class, $classError);
-    }
-
-    /**
-     * Adds a single password field.
-     *
-     * @param string $name       The name of the field.
-     * @param string $value      The value for the field.
-     * @param int    $maxLength  The maximum length for the field.
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @param bool   $HTML       Will the field contain HTML?
-     * @return \SpoonFormPassword
-     */
-    public function addPassword(
-        $name,
-        $value = null,
-        $maxLength = null,
-        $class = null,
-        $classError = null,
-        $HTML = false
-    ) {
-        $name = (string) $name;
-        $value = ($value !== null) ? (string) $value : null;
-        $maxLength = ($maxLength !== null) ? (int) $maxLength : null;
-        $class = ($class !== null) ? (string) $class : 'inputText inputPassword';
-        $classError = ($classError !== null) ? (string) $classError : 'inputTextError inputPasswordError';
-        $HTML = (bool) $HTML;
-
-        // create and return a password field
-        return parent::addPassword($name, $value, $maxLength, $class, $classError, $HTML);
-    }
-
-    /**
-     * Adds a single radio button.
-     *
-     * @param string $name       The name of the element.
-     * @param array  $values     The possible values for the radio button.
-     * @param string $checked    Should the element be checked?
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @return \SpoonFormRadiobutton
-     */
-    public function addRadiobutton($name, array $values, $checked = null, $class = null, $classError = null)
-    {
-        $name = (string) $name;
-        $values = (array) $values;
-        $checked = ($checked !== null) ? (string) $checked : null;
-        $class = ($class !== null) ? (string) $class : 'inputRadio';
-        $classError = ($classError !== null) ? (string) $classError : 'inputRadioError';
-
-        // create and return a radio button
-        return parent::addRadiobutton($name, $values, $checked, $class, $classError);
-    }
-
-    /**
-     * Adds a single textfield.
-     *
-     * @param string $name       The name of the element.
-     * @param string $value      The value inside the element.
-     * @param int    $maxLength  The maximum length for the value.
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @param bool   $HTML       Will this element contain HTML?
-     * @return \SpoonFormText
-     */
-    public function addText($name, $value = null, $maxLength = 255, $class = null, $classError = null, $HTML = false)
-    {
-        $name = (string) $name;
-        $value = ($value !== null) ? (string) $value : null;
-        $maxLength = ($maxLength !== null) ? (int) $maxLength : null;
-        $class = ($class !== null) ? (string) $class : 'inputText';
-        $classError = ($classError !== null) ? (string) $classError : 'inputTextError';
-        $HTML = (bool) $HTML;
-
-        // create and return a textfield
-        return parent::addText($name, $value, $maxLength, $class, $classError, $HTML);
-    }
-
-    /**
-     * Adds a single textarea.
-     *
-     * @param string $name       The name of the element.
-     * @param string $value      The value inside the element.
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @param bool   $HTML       Will the element contain HTML?
-     * @return \SpoonFormTextarea
-     */
-    public function addTextarea($name, $value = null, $class = null, $classError = null, $HTML = false)
-    {
-        $name = (string) $name;
-        $value = ($value !== null) ? (string) $value : null;
-        $class = ($class !== null) ? (string) $class : 'textarea';
-        $classError = ($classError !== null) ? (string) $classError : 'textareaError';
-        $HTML = (bool) $HTML;
-
-        // create and return a textarea
-        return parent::addTextarea($name, $value, $class, $classError, $HTML);
-    }
-
-    /**
-     * Adds a single time field.
-     *
-     * @param string $name       The name of the element.
-     * @param string $value      The value inside the element.
-     * @param string $class      Class(es) that will be applied on the element.
-     * @param string $classError Class(es) that will be applied on the element when an error occurs.
-     * @return \SpoonFormTime
-     */
-    public function addTime($name, $value = null, $class = null, $classError = null)
-    {
-        $name = (string) $name;
-        $value = ($value !== null) ? (string) $value : null;
-        $class = ($class !== null) ? (string) $class : 'inputText inputTime';
-        $classError = ($classError !== null) ? (string) $classError : 'inputTextError inputTimeError';
-
-        // create and return a time field
-        return parent::addTime($name, $value, $class, $classError);
     }
 
     /**
